+++ src/main/java/org/realityforge/vecmath/Matrix4d.java
@@ -494,6 +494,106 @@ public final class Matrix4d
                 0, 0, ( zFar + zNear ) / ( zNear - zFar ), ( zFar + zFar ) * zNear / ( zNear - zFar ),
                 0, 0, -1, 0 );
   }
+
+  /**
+   * Apply a "lookat" transformation to this matrix for a right-handed coordinate system,
+   * that aligns <code>-z</code> with <code>center - eye</code>.
+   *
+   * @param eye    the position of the camera
+   * @param center the point in space to look at
+   * @param up     the direction of 'up'
+   * @return this
+   * @see <a href="https://github.com/JOML-CI/JOML/blob/main/src/org/joml/Matrix4d.java">Derived from equivalent in org.joml.Matrix4d</a>
+   */
+  @Nonnull
+  public Matrix4d lookAt( @Nonnull final Vector3d eye, @Nonnull final Vector3d center, @Nonnull final Vector3d up )
+  {
+    return lookAt( eye.x, eye.y, eye.z, center.x, center.y, center.z, up.x, up.y, up.z );
+  }
+
+  /**
+   * Apply a "lookat" transformation to this matrix for a right-handed coordinate system,
+   * that aligns <code>-z</code> with <code>center - eye</code>.
+   *
+   * @param eyeX    the x-coordinate of the eye/camera location
+   * @param eyeY    the y-coordinate of the eye/camera location
+   * @param eyeZ    the z-coordinate of the eye/camera location
+   * @param centerX the x-coordinate of the point to look at
+   * @param centerY the y-coordinate of the point to look at
+   * @param centerZ the z-coordinate of the point to look at
+   * @param upX     the x-coordinate of the up vector
+   * @param upY     the y-coordinate of the up vector
+   * @param upZ     the z-coordinate of the up vector
+   * @return this
+   * @see <a href="https://github.com/JOML-CI/JOML/blob/main/src/org/joml/Matrix4d.java">Derived from equivalent in org.joml.Matrix4d</a>
+   */
+  @SuppressWarnings( "UnnecessaryLocalVariable" )
+  @Nonnull
+  public Matrix4d lookAt( final double eyeX,
+                          final double eyeY,
+                          final double eyeZ,
+                          final double centerX,
+                          final double centerY,
+                          final double centerZ,
+                          final double upX,
+                          final double upY,
+                          final double upZ )
+  {
+    double dirX = eyeX - centerX;
+    double dirY = eyeY - centerY;
+    double dirZ = eyeZ - centerZ;
+    // Normalize direction
+    final double invDirLength = 1.0 / Math.sqrt( dirX * dirX + dirY * dirY + dirZ * dirZ );
+    dirX *= invDirLength;
+    dirY *= invDirLength;
+    dirZ *= invDirLength;
+    // left = up x direction
+    double leftX = upY * dirZ - upZ * dirY;
+    double leftY = upZ * dirX - upX * dirZ;
+    double leftZ = upX * dirY - upY * dirX;
+    // normalize left
+    double invLeftLength = 1.0 / Math.sqrt( leftX * leftX + leftY * leftY + leftZ * leftZ );
+    leftX *= invLeftLength;
+    leftY *= invLeftLength;
+    leftZ *= invLeftLength;
+    // up = direction x left
+    final double upnX = dirY * leftZ - dirZ * leftY;
+    final double upnY = dirZ * leftX - dirX * leftZ;
+    final double upnZ = dirX * leftY - dirY * leftX;
+    // calculate right matrix elements
+    final double rm00 = leftX;
+    final double rm01 = upnX;
+    final double rm02 = dirX;
+    final double rm10 = leftY;
+    final double rm11 = upnY;
+    final double rm12 = dirY;
+    final double rm20 = leftZ;
+    final double rm21 = upnZ;
+    final double rm22 = dirZ;
+    final double rm30 = -( leftX * eyeX + leftY * eyeY + leftZ * eyeZ );
+    final double rm31 = -( upnX * eyeX + upnY * eyeY + upnZ * eyeZ );
+    final double rm32 = -( dirX * eyeX + dirY * eyeY + dirZ * eyeZ );
+    final double nm00 = m00 * rm00 + m10 * rm01 + m20 * rm02;
+    final double nm01 = m01 * rm00 + m11 * rm01 + m21 * rm02;
+    final double nm02 = m02 * rm00 + m12 * rm01 + m22 * rm02;
+    final double nm03 = m03 * rm00 + m13 * rm01 + m23 * rm02;
+    final double nm10 = m00 * rm10 + m10 * rm11 + m20 * rm12;
+    final double nm11 = m01 * rm10 + m11 * rm11 + m21 * rm12;
+    final double nm12 = m02 * rm10 + m12 * rm11 + m22 * rm12;
+    final double nm13 = m03 * rm10 + m13 * rm11 + m23 * rm12;
+    final double nm20 = m00 * rm20 + m10 * rm21 + m20 * rm22;
+    final double nm21 = m01 * rm20 + m11 * rm21 + m21 * rm22;
+    final double nm22 = m02 * rm20 + m12 * rm21 + m22 * rm22;
+    final double nm23 = m03 * rm20 + m13 * rm21 + m23 * rm22;
+    final double nm30 = m00 * rm30 + m10 * rm31 + m20 * rm32 + m30;
+    final double nm31 = m01 * rm30 + m11 * rm31 + m21 * rm32 + m31;
+    final double nm32 = m02 * rm30 + m12 * rm31 + m22 * rm32 + m32;
+    final double nm33 = m03 * rm30 + m13 * rm31 + m23 * rm32 + m33;
+    return set( nm00, nm10, nm20, nm30,
+                nm01, nm11, nm21, nm31,
+                nm02, nm12, nm22, nm32,
+                nm03, nm13, nm23, nm33 );
+  }
   /**
    * Fill the specified target with the matrix components starting at the specified offset.
    * The components are emitted in column-major form.